requirements.txt
awsebcli==3.7.6
blessed==1.9.5
boto==2.40.0
botocore==1.4.15
cement==2.8.2
colorama==0.3.7
Django==1.9.5
django-debug-toolbar==1.4
django-storages==1.4.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
pathspec==0.3.4
python-dateutil==2.5.3
PyYAML==3.11
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
sqlparse==0.1.19
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.37.0
